+++ wemake_python_styleguide/visitors/ast/comparisons.py
@@ -5,14 +5,16 @@ from typing import ClassVar, Sequence
 
 from wemake_python_styleguide.logics.nodes import is_literal
 from wemake_python_styleguide.logics.variables import is_same_variable
-from wemake_python_styleguide.types import AnyNodes, final
+from wemake_python_styleguide.types import AnyIf, AnyNodes, final
 from wemake_python_styleguide.violations.consistency import (
     ComparisonOrderViolation,
     ConstantComparisonViolation,
     MultipleInComparisonViolation,
     RedundantComparisonViolation,
+    WrongConditionalViolation,
 )
 from wemake_python_styleguide.visitors.base import BaseNodeVisitor
+from wemake_python_styleguide.visitors.decorators import alias
 
 
 @final
@@ -138,3 +140,35 @@ class WrongComparisionOrderVisitor(BaseNodeVisitor):
         """
         self._check_ordering(node)
         self.generic_visit(node)
+
+
+@alias('visit_any_if', (
+    'visit_If',
+    'visit_IfExp',
+))
+class WrongConditionalVisitor(BaseNodeVisitor):
+    """Finds wrong conditional arguments."""
+
+    _forbidden_nodes: ClassVar[AnyNodes] = (
+        ast.List,
+        ast.Set,
+        ast.Num,
+        ast.NameConstant,
+        ast.Str,
+        ast.Dict,
+    )
+
+    def visit_any_if(self, node: AnyIf) -> None:
+        """
+        Ensures that if statements are using valid conditionals.
+
+        Raises:
+            WrongConditionalViolation
+
+        """
+        self._check_if_statement_conditional(node)
+        self.generic_visit(node)
+
+    def _check_if_statement_conditional(self, node: AnyIf) -> None:
+        if isinstance(node.test, self._forbidden_nodes):
+            self.add_violation(WrongConditionalViolation(node))